Remove nesting from IWaitForChangedResult
