+++ Source/Testably.Abstractions.Interface/FileSystem/IFileSystemWatcher.cs
@@ -75,22 +75,4 @@ public interface IFileSystemWatcher : IFileSystemEntity, IDisposable
 	/// <inheritdoc cref="FileSystemWatcher.WaitForChanged(WatcherChangeTypes, TimeSpan)" />
 	IWaitForChangedResult WaitForChanged(WatcherChangeTypes changeType, TimeSpan timeout);
 #endif
-
-	/// <summary>
-	///     Abstractions for <see cref="WaitForChangedResult" />.
-	/// </summary>
-	public interface IWaitForChangedResult
-	{
-		/// <inheritdoc cref="WaitForChangedResult.ChangeType" />
-		WatcherChangeTypes ChangeType { get; }
-
-		/// <inheritdoc cref="WaitForChangedResult.Name" />
-		string? Name { get; }
-
-		/// <inheritdoc cref="WaitForChangedResult.OldName" />
-		string? OldName { get; }
-
-		/// <inheritdoc cref="WaitForChangedResult.TimedOut" />
-		bool TimedOut { get; }
-	}
 }
