+++ Source/Testably.Abstractions.Interface/FileSystem/IWaitForChangedResult.cs
@@ -1,0 +1,21 @@
+﻿using System.IO;
+
+namespace Testably.Abstractions.FileSystem;
+
+/// <summary>
+///     Abstractions for <see cref="WaitForChangedResult" />.
+/// </summary>
+public interface IWaitForChangedResult
+{
+	/// <inheritdoc cref="WaitForChangedResult.ChangeType" />
+	WatcherChangeTypes ChangeType { get; }
+
+	/// <inheritdoc cref="WaitForChangedResult.Name" />
+	string? Name { get; }
+
+	/// <inheritdoc cref="WaitForChangedResult.OldName" />
+	string? OldName { get; }
+
+	/// <inheritdoc cref="WaitForChangedResult.TimedOut" />
+	bool TimedOut { get; }
+}
