+++ Tests/Testably.Abstractions.Tests/FileSystem/FileSystemWatcher/Tests.cs
@@ -37,7 +37,7 @@ public abstract partial class Tests<TFileSystem>
 					FileSystem.Directory.Delete(path);
 				}
 			});
-			IFileSystemWatcher.IWaitForChangedResult result =
+			IWaitForChangedResult result =
 				fileSystemWatcher.WaitForChanged(WatcherChangeTypes.Created, 1000);
 
 			fileSystemWatcher.EnableRaisingEvents.Should().BeTrue();
@@ -89,7 +89,7 @@ public abstract partial class Tests<TFileSystem>
 					FileSystem.Directory.Delete(path);
 				}
 			});
-			IFileSystemWatcher.IWaitForChangedResult result =
+			IWaitForChangedResult result =
 				fileSystemWatcher.WaitForChanged(WatcherChangeTypes.Created, 100);
 
 			fileSystemWatcher.EnableRaisingEvents.Should().BeTrue();
